# Optimization for String Encoding/Decoding