@xinchen10 2019-02-27
I did a quick test for write. The result is not as comprehensive as yours but it does give something for different APIs. The unsafe overload is about 1% better than the pure managed one. Both are significantly better than the one with extra allocation. I didn't run read tests but I would expect similar results.

```
Size 16 iterations 50000000 string 'sEqb3GRO...'
Write1: 57346704
Write2: 20605029
Write3: 20284336
Size 128 iterations 50000000 string '2B31JuM8...'
Write1: 153184622
Write2: 63810829
Write3: 63509337
Size 1024 iterations 50000000 string '3iReWaRm...'
Write1: 916739030
Write2: 484157501
Write3: 441844717
```

The code,
```
    class Program
    {
        static Random random = new Random();

        static void Main(string[] args)
        {
            int[] size = new[] { 16, 128, 1024 };
            byte[] buffer = new byte[8192];
            long iterations = 50000000;
            Stopwatch w = Stopwatch.StartNew();

            for (int t = 0; t < size.Length; t++)
            {
                string test = GetTestString(size[t]);
                Console.WriteLine($"Size {size[t]} iterations {iterations} string '{test.Substring(0, 8)}...'");

                w.Restart();
                for (long i = 0; i < iterations; i++)
                {
                    Write1(test, buffer);
                }
                w.Stop();
                Console.WriteLine($"Write1: {w.ElapsedTicks}");

                w.Restart();
                for (long i = 0; i < iterations; i++)
                {
                    Write2(test, buffer);
                }
                w.Stop();
                Console.WriteLine($"Write2: {w.ElapsedTicks}");

                w.Restart();
                for (long i = 0; i < iterations; i++)
                {
                    Write3(test, buffer);
                }
                w.Stop();
                Console.WriteLine($"Write3: {w.ElapsedTicks}");
            }
        }

        static string GetTestString(int len)
        {
            const string chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz012345678";
            return new string(Enumerable.Range(0, len).Select(i => chars[random.Next(chars.Length)]).ToArray());
        }

        static int Write1(string s, byte[] buffer)
        {
            byte[] data = Encoding.UTF8.GetBytes(s);
            Buffer.BlockCopy(data, 0, buffer, 4, data.Length);
            return data.Length;
        }

        static int Write2(string s, byte[] buffer)
        {
            return Encoding.UTF8.GetBytes(s, 0, s.Length, buffer, 0);
        }

        static unsafe int Write3(string s, byte[] buffer)
        {
            fixed (byte* b = buffer)
            {
                fixed (char* p = s)
                {
                    return Encoding.UTF8.GetBytes(p, s.Length, b, buffer.Length);
                }
            }
        }

        static string Read1(byte[] buffer, int count)
        {
            return new string(Encoding.UTF8.GetChars(buffer, 0, count));
        }

        static string Read2(byte[] buffer, int count)
        {
            return Encoding.UTF8.GetString(buffer, 0, count);
        }

        static unsafe string Read3(int len, byte[] buffer, int count)
        {
            char* c = stackalloc char[len];
            fixed (byte* p = buffer)
            {
                Encoding.UTF8.GetChars(p, count, c, len);
                return new string(c, 0, len);
            }
        }
    }
```