@szehetner 2019-02-27
One problem with the GetBytes(String, Int32, Int32, Byte[], Int32) overload is that there is no way to get the byte count before writing the bytes into the ByteBuffer. This complicates the writing of the string "header" with the byte count. This either needs a seperate Encoding.GetByteCount() call or a possible copy afterwards to accomodate smallEncoding with a larger byte count, both options would sure be slower.

The implementation of the GetBytes(String, Int32, Int32, Byte[], Int32) overload anyway looks pretty much the same as the version with the char pointer, at least for .NET Framework 4.7.2 (see https://referencesource.microsoft.com/#mscorlib/system/text/utf8encoding.cs,200). 

So I would suggest to use the byte pointer overload in this case.