@xinchen10 2019-02-27
We need to call GetByteCount even for char* overload. I am trying to avoid unsafe code so if someone has partially trusted or transparent code they can still use it. The only place we currently use unsafe is for reading/writing float and double types. I am trying to remove that as well.