@szehetner 2019-02-27
I changed it to the safe GetBytes overload. It is a bit slower than the unsafe version, but still much faster than the old version and with zero allocations.

```
               Method | StringLength |        Mean |      Error |     StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
--------------------- |------------- |------------:|-----------:|-----------:|------------:|------------:|------------:|--------------------:|
      WriteString_Old |            1 |    91.53 ns |  1.8752 ns |  3.8305 ns |      0.0101 |           - |           - |                16 B |
  WriteString_Pointer |            1 |    42.09 ns |  0.3434 ns |  0.2681 ns |           - |           - |           - |                   - |
 WriteString_GetBytes |            1 |    42.91 ns |  1.4489 ns |  1.3553 ns |           - |           - |           - |                   - |
      WriteString_Old |          128 |   284.66 ns |  6.4430 ns |  7.9126 ns |      0.0887 |           - |           - |               140 B |
  WriteString_Pointer |          128 |   211.59 ns |  5.8843 ns |  5.2162 ns |           - |           - |           - |                   - |
 WriteString_GetBytes |          128 |   220.24 ns |  1.5534 ns |  1.4531 ns |           - |           - |           - |                   - |
      WriteString_Old |         1024 | 1,594.45 ns | 15.8474 ns | 14.0483 ns |      0.6580 |           - |           - |              1036 B |
  WriteString_Pointer |         1024 | 1,352.71 ns |  7.1662 ns |  6.3527 ns |           - |           - |           - |                   - |
 WriteString_GetBytes |         1024 | 1,370.63 ns | 39.3753 ns | 34.9052 ns |           - |           - |           - |                   - |