@xinchen10 2019-02-27
This small difference is acceptable. Thanks for optimizing the code and measuring the result!